+++ src/main/java/com/powsybl/benchmark/commons/Constants.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.benchmark;
+package com.powsybl.benchmark.commons;
 
 /**
  * @author Nicolas Rol {@literal <nicolas.rol at rte-france.com>}
