+++ src/main/java/com/powsybl/benchmark/loadflow/state/Rte1888NetworkState.java
@@ -5,9 +5,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.benchmark.state;
+package com.powsybl.benchmark.loadflow.state;
 
-import static com.powsybl.benchmark.Constants.RTE_1888;
+import static com.powsybl.benchmark.commons.Constants.RTE_1888;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
