+++ src/main/java/com/powsybl/benchmark/security/AbstractSecurityAnalysisBenchmark.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.benchmark;
+package com.powsybl.benchmark.security;
 
 import com.powsybl.contingency.Contingency;
 import com.powsybl.iidm.network.Network;
