+++ src/main/java/com/powsybl/benchmark/security/ManualSecurityAnalysisBenchmark.java
@@ -1,14 +1,15 @@
-/**
- * Copyright (c) 2022, RTE (http://www.rte-france.com)
+/*
+ * Copyright (c) 2022-2026, RTE (https://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.benchmark;
+package com.powsybl.benchmark.security;
 
 import com.google.common.base.Stopwatch;
-import com.powsybl.benchmark.state.LoadFlowParametersType;
-import com.powsybl.benchmark.state.MatpowerUtil;
+import com.powsybl.benchmark.loadflow.state.LoadFlowParametersType;
+import com.powsybl.benchmark.commons.MatpowerUtil;
 import com.powsybl.contingency.Contingency;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.security.SecurityAnalysis;
@@ -22,8 +23,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import static com.powsybl.benchmark.Constants.RTE_1888;
-import static com.powsybl.benchmark.Constants.RTE_6515;
+import static com.powsybl.benchmark.commons.Constants.RTE_1888;
+import static com.powsybl.benchmark.commons.Constants.RTE_6515;
 
 /**
  * @deprecated since 2025.4.0
